"backend changes and refactored code with riverpod and dio"

+++ finance-tracker-backend/src/main/java/com/finance/tracker/transactions/domain/CategoryExpenseSummary.java
@@ -1,9 +1,11 @@
 package com.finance.tracker.transactions.domain;
 
+import java.math.BigDecimal;
+
 public record CategoryExpenseSummary (
 
         String categoryId,
         String categoryName,
-        Double total,
+        BigDecimal total,
         Long transactionCount
 ){}
